+++ workflow-trace-viewer/src/jvmMain/kotlin/com/squareup/workflow1/traceviewer/ui/control/SearchBox.kt
@@ -1,4 +1,4 @@
-package com.squareup.workflow1.traceviewer.ui
+package com.squareup.workflow1.traceviewer.ui.control
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
