+++ workflow-trace-viewer/src/jvmMain/kotlin/com/squareup/workflow1/traceviewer/util/parser/JsonParser.kt
@@ -1,4 +1,4 @@
-package com.squareup.workflow1.traceviewer.util
+package com.squareup.workflow1.traceviewer.util.parser
 
 import com.squareup.moshi.JsonAdapter
 import com.squareup.moshi.Moshi
@@ -7,6 +7,8 @@ import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import com.squareup.workflow1.traceviewer.model.Node
 import com.squareup.workflow1.traceviewer.model.addChild
 import com.squareup.workflow1.traceviewer.model.replaceChild
+import com.squareup.workflow1.traceviewer.util.parser.ParseResult.Failure
+import com.squareup.workflow1.traceviewer.util.parser.ParseResult.Success
 import io.github.vinceglb.filekit.PlatformFile
 import io.github.vinceglb.filekit.readString
 import kotlin.reflect.jvm.javaType
@@ -31,11 +33,11 @@ internal suspend fun parseFileTrace(
   val jsonString = file.readString()
   val workflowAdapter = createMoshiAdapter<List<Node>>()
   val parsedRenderPasses = try {
-    workflowAdapter.fromJson(jsonString) ?: return ParseResult.Failure(
+    workflowAdapter.fromJson(jsonString) ?: return Failure(
       IllegalArgumentException("Provided trace file is empty or malformed.")
     )
   } catch (e: Exception) {
-    return ParseResult.Failure(e)
+    return Failure(e)
   }
 
   val parsedFrames = parsedRenderPasses.map { renderPass -> getFrameFromRenderPass(renderPass) }
@@ -45,7 +47,7 @@ internal suspend fun parseFileTrace(
     frameTrees.add(mergedTree)
     mergedTree
   }
-  return ParseResult.Success(
+  return Success(
     trace = parsedFrames,
     trees = frameTrees,
     affectedNodes = parsedRenderPasses
@@ -64,11 +66,11 @@ internal fun parseLiveTrace(
   currentTree: Node? = null
 ): ParseResult {
   val parsedRenderPass = try {
-    adapter.fromJson(renderPass) ?: return ParseResult.Failure(
+    adapter.fromJson(renderPass) ?: return Failure(
       IllegalArgumentException("Provided trace data is empty or malformed.")
     )
   } catch (e: Exception) {
-    return ParseResult.Failure(e)
+    return Failure(e)
   }
 
   val parsedFrame = getFrameFromRenderPass(parsedRenderPass)
@@ -81,7 +83,7 @@ internal fun parseLiveTrace(
   }
 
   // Since live tracing handles one frame at a time, we generalize and return listOf for each.
-  return ParseResult.Success(
+  return Success(
     trace = listOf(parsedFrame),
     trees = listOf(mergedTree),
     affectedNodes = listOf(parsedRenderPass)
@@ -153,7 +155,8 @@ internal fun mergeFrameIntoMainTree(
 }
 
 internal sealed interface ParseResult {
-  class Success(val trace: List<Node>, val trees: List<Node>, affectedNodes: List<List<Node>>) : ParseResult {
+  class Success(val trace: List<Node>, val trees: List<Node>, affectedNodes: List<List<Node>>) :
+    ParseResult {
     val affectedNodes = affectedNodes.map { it.toSet() }
   }
   class Failure(val error: Throwable) : ParseResult
